# openapi.yaml
...
openapi: 3.1.0
info:
  title: Issues API Testbed
  version: 0.1.0
  summary: A validated wrapper over the GitHub Issues REST API, with signed webhook ingestion.
  description: |
    This service exposes its own HTTP API for issue CRUD and comments on a single
    GitHub repository, and ingests `issues` / `issue_comment` webhooks from that
    repository.

    ### Authentication
    Clients of *this* service do not authenticate. The GitHub credential is held
    server-side in the `GITHUB_TOKEN` environment variable and is never accepted
    from, or echoed to, a caller. The `githubToken` security scheme below documents
    that server-to-GitHub relationship so the dependency is visible in the contract.

    `POST /webhook` is authenticated differently: GitHub signs each delivery with
    `WEBHOOK_SECRET`, and the service verifies the `X-Hub-Signature-256` header with
    a constant-time comparison over the raw request body.

    ### Errors
    Every failure — validation, upstream, or unexpected — uses the same `Error`
    envelope, so a client parses one shape. `details` is free-form and carries the
    upstream context when there is any.
  contact:
    name: Anita Agasaveeran
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: "8000"
...
paths:
...
  /issues:
    post:
      tags: [issues]
      operationId: createIssue
      summary: Create an issue
      description: Creates an issue in the configured repository.
      security:
        - githubToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCreate"
            examples:
              minimal:
                summary: Title only
                value:
                  title: Checkout button is dead
              labelled:
                summary: With body and labels
                value:
                  title: Checkout button is dead
                  body: Clicking it does nothing on Safari 17.
                  labels: [bug, ops]
      responses:
        "201":
          description: The issue was created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
              examples:
                created:
                  $ref: "#/components/examples/IssueCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"

    get:
      tags: [issues]
      operationId: listIssues
      summary: List issues
      description: |
        Lists issues, preserving GitHub's pagination semantics. Pull requests are
        filtered out: GitHub's issues endpoint returns them, this service does not.

        The upstream `Link` header is rewritten to point at this service's own
        `/issues` route, keeping only `page` and `per_page`.

        Two layers of conditional GET are at work. Sending `If-None-Match` with a
        previously returned `ETag` performs a conditional GET and is answered with
        `304`. Independently, the service caches the ETag of each distinct query and
        replays it upstream on the next identical request; when GitHub answers `304`
        the cached body is returned as a normal `200`, so the caller sees no
        difference but the request consumed no GitHub rate-limit quota. The `X-Cache`
        response header reports which path was taken.
      security:
        - githubToken: []
      parameters:
        - $ref: "#/components/parameters/StateFilter"
        - $ref: "#/components/parameters/Labels"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: A page of issues.
          headers:
            Link:
              $ref: "#/components/headers/Link"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Page:
              $ref: "#/components/headers/XPage"
            X-Per-Page:
              $ref: "#/components/headers/XPerPage"
            X-Cache:
              $ref: "#/components/headers/XCache"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
              examples:
                page:
                  summary: One open issue
                  value:
                    - number: 42
                      title: Checkout button is dead
                      body: Clicking it does nothing on Safari 17.
                      state: open
                      labels: [bug, ops]
                      html_url: https://github.com/anita-agasaveeran/issues-api-testbed/issues/42
                      user: anita-agasaveeran
                      comments: 3
                      created_at: "2026-09-01T10:00:00Z"
                      updated_at: "2026-09-02T11:30:00Z"
                      closed_at: null
                empty:
                  summary: No matching issues
                  value: []
        "304":
          description: |
            Not modified. The caller's `If-None-Match` matched the current upstream
            `ETag`, so no body is returned and no GitHub quota was consumed.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /issues/{number}:
    parameters:
      - $ref: "#/components/parameters/IssueNumber"
    get:
      tags: [issues]
      operationId: getIssue
      summary: Get a single issue
      description: |
        Returns one issue. A number belonging to a pull request is reported as
        `404`, even though GitHub would return it from this endpoint.
      security:
        - githubToken: []
      responses:
        "200":
          description: The issue.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
              examples:
                found:
                  $ref: "#/components/examples/IssueCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"

    patch:
      tags: [issues]
      operationId: updateIssue
      summary: Update an issue
      description: |
        Renames an issue, edits its body, or closes and reopens it. At least one of
        `title`, `body`, or `state` must be supplied; an empty object is a `400`
        rather than a no-op round trip to GitHub.
      security:
        - githubToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssuePatch"
            examples:
              rename:
                summary: Rename
                value:
                  title: Checkout button unresponsive on Safari
              close:
                summary: Close
                value:
                  state: closed
              reopen:
                summary: Reopen
                value:
                  state: open
      responses:
        "200":
          description: The updated issue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
              examples:
                closed:
                  summary: After closing
                  value:
                    number: 42
                    title: Checkout button is dead
                    body: Clicking it does nothing on Safari 17.
                    state: closed
                    labels: [bug, ops]
                    html_url: https://github.com/anita-agasaveeran/issues-api-testbed/issues/42
                    user: anita-agasaveeran
                    comments: 3
                    created_at: "2026-09-01T10:00:00Z"
                    updated_at: "2026-09-03T08:00:00Z"
                    closed_at: "2026-09-03T08:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /issues/{number}/comments:
    parameters:
      - $ref: "#/components/parameters/IssueNumber"
    post:
      tags: [issues]
      operationId: createComment
      summary: Comment on an issue
      security:
        - githubToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreate"
            examples:
              comment:
                summary: A plain comment
                value:
                  body: Reproduced on Safari 17.
      responses:
        "201":
          description: The comment was created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
              examples:
                created:
                  $ref: "#/components/examples/CommentCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"

    get:
      tags: [issues]
      operationId: listComments
      summary: List an issue's comments
      security:
        - githubToken: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: A page of comments.
          headers:
            Link:
              $ref: "#/components/headers/Link"
            X-Page:
              $ref: "#/components/headers/XPage"
            X-Per-Page:
              $ref: "#/components/headers/XPerPage"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
              examples:
                page:
                  summary: One comment
                  value:
                    - id: 3100000001
                      body: Reproduced on Safari 17.
                      user: anita-agasaveeran
                      html_url: https://github.com/anita-agasaveeran/issues-api-testbed/issues/42#issuecomment-3100000001
                      created_at: "2026-09-02T12:00:00Z"
                      updated_at: "2026-09-02T12:00:00Z"
                      issue_number: 42
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
...
components:
  securitySchemes:
    githubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub personal access token
      description: |
        The token this service presents **to GitHub**, read from `GITHUB_TOKEN` at
        startup. It is never supplied by a client of this API and never appears in
        a response or a log line. Requires `Issues: read & write` and
        `Metadata: read` on the configured repository.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: |
        HMAC-SHA256 of the raw request body, keyed with `WEBHOOK_SECRET` and sent by
        GitHub as `sha256=<hex digest>`. Verified with a constant-time comparison.
  parameters:
    IssueNumber:
      name: number
      in: path
      required: true
      description: The issue number within the repository.
      schema:
        type: integer
        minimum: 1
      example: 42
    StateFilter:
      name: state
      in: query
      required: false
      description: Filter by issue state.
      schema:
        $ref: "#/components/schemas/IssueStateFilter"
      example: open
    Labels:
      name: labels
      in: query
      required: false
      description: Comma-separated label names; an issue must carry all of them.
      schema:
        type: string
      example: bug,ops
    Page:
      name: page
      in: query
      required: false
      description: 1-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      required: false
      description: Results per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: An `ETag` from a previous list response, for a conditional GET.
      schema:
        type: string
      example: W/"6c1b2f0e9a"
  headers:
    Location:
      description: Path of the newly created resource on this service.
      schema:
        type: string
      example: /issues/42
    Link:
      description: |
        RFC 8288 pagination links, rewritten to point at this service. Only `page`
        and `per_page` are carried over from upstream.
      schema:
        type: string
      example: '</issues?page=2&per_page=30>; rel="next", </issues?page=5&per_page=30>; rel="last"'
    ETag:
      description: Upstream entity tag, usable as `If-None-Match` on a later request.
      schema:
        type: string
      example: W/"6c1b2f0e9a"
    XPage:
      description: The page number that produced this response.
      schema:
        type: integer
      example: 1
    XPerPage:
      description: The page size that produced this response.
      schema:
        type: integer
      example: 30
    XCache:
      description: |
        Whether this response was served from the ETag cache. `HIT` means GitHub
        answered `304` and the cached body was replayed, costing no rate-limit quota.
      schema:
        type: string
        enum: [HIT, MISS]
      example: HIT
    RetryAfter:
      description: Seconds to wait before retrying, derived from GitHub's rate-limit headers.
      schema:
        type: integer
        minimum: 1
      example: 45
    RequestId:
      description: Correlation id, echoed from the request or generated per request.
      schema:
        type: string
      example: b00d0e4c-46da-4bc3-8ec6-66d22f349b45
  schemas:
    IssueState:
      type: string
      enum: [open, closed]
      description: A state an issue can be set to.
    IssueStateFilter:
      type: string
      enum: [open, closed, all]
      default: open
      description: A state that issues can be filtered on.
    IssueCreate:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 256
          description: Leading and trailing whitespace is stripped; a blank title is rejected.
          examples: ["Checkout button is dead"]
        body:
          type: [string, "null"]
          maxLength: 65536
          examples: ["Clicking it does nothing on Safari 17."]
        labels:
          type: [array, "null"]
          maxItems: 100
          items:
            type: string
            minLength: 1
          description: Label names. Blank entries are rejected rather than silently dropped.
          examples: [[bug, ops]]
    IssuePatch:
      type: object
      additionalProperties: false
      minProperties: 1
      description: At least one of `title`, `body`, or `state` must be present.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 256
        body:
          type: string
          maxLength: 65536
        state:
          $ref: "#/components/schemas/IssueState"
    CommentCreate:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 65536
          examples: ["Reproduced on Safari 17."]
    Issue:
      type: object
      required: [number, title, state, labels, html_url, comments, created_at, updated_at]
      description: |
        A projection of GitHub's issue object. Labels and the author are flattened
        to plain strings so that request and response shapes agree.
      properties:
        number:
          type: integer
          minimum: 1
          examples: [42]
        title:
          type: string
          examples: ["Checkout button is dead"]
        body:
          type: [string, "null"]
        state:
          type: string
          examples: [open]
        labels:
          type: array
          items:
            type: string
          examples: [[bug, ops]]
        html_url:
          type: string
          format: uri
        user:
          type: [string, "null"]
          description: Login of the issue's author.
        comments:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: [string, "null"]
          format: date-time
    Comment:
      type: object
      required: [id, body, html_url, created_at]
      properties:
        id:
          type: integer
          examples: [3100000001]
        body:
          type: string
        user:
          type: [string, "null"]
        html_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: [string, "null"]
          format: date-time
        issue_number:
          type: [integer, "null"]
          description: Echoed from the request path for convenience.
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          enum:
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - method_not_allowed
            - rate_limited
            - upstream_error
            - upstream_unavailable
            - internal_error
            - http_error
        message:
          type: string
          description: Human-readable explanation, safe to surface to an operator.
        details:
          description: |
            Free-form context. Carries `upstream_status` and `upstream_message` when
            the failure came from GitHub. Never contains secrets or signatures.
        request_id:
          type: [string, "null"]
          description: Correlation id, matching the `X-Request-ID` response header.
    Error:
      type: object
      required: [error]
      description: The single error envelope used by every failing response.
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
  examples:
    IssueCreated:
      summary: A newly created issue
      value:
        number: 42
        title: Checkout button is dead
        body: Clicking it does nothing on Safari 17.
        state: open
        labels: [bug, ops]
        html_url: https://github.com/anita-agasaveeran/issues-api-testbed/issues/42
        user: anita-agasaveeran
        comments: 0
        created_at: "2026-09-01T10:00:00Z"
        updated_at: "2026-09-01T10:00:00Z"
        closed_at: null
    CommentCreated:
      summary: A newly created comment
      value:
        id: 3100000001
        body: Reproduced on Safari 17.
        user: anita-agasaveeran
        html_url: https://github.com/anita-agasaveeran/issues-api-testbed/issues/42#issuecomment-3100000001
        created_at: "2026-09-02T12:00:00Z"
        updated_at: "2026-09-02T12:00:00Z"
        issue_number: 42
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingTitle:
              summary: Missing required field
              value:
                error:
                  code: validation_error
                  message: Request payload or query parameters are invalid.
                  details:
                    - location: [body, title]
                      message: Field required
                      type: missing
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
            invalidState:
              summary: Unsupported state filter
              value:
                error:
                  code: validation_error
                  message: Request payload or query parameters are invalid.
                  details:
                    - location: [query, state]
                      message: "Input should be 'open', 'closed' or 'all'"
                      type: enum
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
    Unauthorized:
      description: GitHub rejected the configured token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            badCredentials:
              summary: Expired or revoked token
              value:
                error:
                  code: unauthorized
                  message: >-
                    GitHub rejected the configured token. Check GITHUB_TOKEN is set,
                    unexpired, and has Issues: read & write on this repository.
                  details:
                    upstream_status: 401
                    upstream_message: Bad credentials
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
    Forbidden:
      description: The token is valid but lacks permission for this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            insufficientScope:
              summary: Token missing the Issues permission
              value:
                error:
                  code: forbidden
                  message: GitHub refused this request. The token is valid but lacks permission for it.
                  details:
                    upstream_status: 403
                    upstream_message: Resource not accessible by personal access token
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
    NotFound:
      description: The issue or repository does not exist, or the token cannot see it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingIssue:
              summary: No such issue
              value:
                error:
                  code: not_found
                  message: Issue #999 does not exist in this repository.
                  details:
                    issue_number: 999
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
    RateLimited:
      description: |
        The GitHub rate limit is exhausted. GitHub signals this with `403` plus
        `x-ratelimit-remaining: 0`; this service translates it to `429` and returns
        a `Retry-After` derived from the reset time.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            quotaExhausted:
              summary: Primary rate limit
              value:
                error:
                  code: rate_limited
                  message: GitHub API rate limit exceeded. Retry in 45s.
                  details:
                    upstream_status: 403
                    upstream_message: API rate limit exceeded
                    retry_after_seconds: 45
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
    UpstreamUnavailable:
      description: GitHub is unreachable, timed out, or returned a server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            upstreamDown:
              summary: GitHub 5xx
              value:
                error:
                  code: upstream_unavailable
                  message: GitHub returned a server error. This is upstream, not your request.
                  details:
                    upstream_status: 502
                    upstream_message: Bad gateway
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
            timeout:
              summary: Timed out
              value:
                error:
                  code: upstream_unavailable
                  message: Timed out talking to the GitHub API. Please retry.
                  details:
                    upstream: timeout
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98